+++ packages/core/src/browser/keyboard/browser-keyboard-layout-provider.ts
@@ -19,10 +19,12 @@ import { IKeyboardLayoutInfo } from 'native-keymap';
 import { isOSX } from '../../common/os';
 import { Emitter } from '../../common/event';
 import { ILogger } from '../../common/logger';
+import { Deferred } from '../../common/promise-util';
 import { NativeKeyboardLayout, KeyboardLayoutProvider, KeyboardLayoutChangeNotifier } from '../../common/keyboard/keyboard-layout-provider';
+import { LocalStorageService } from '../storage-service';
 import { KeyCode } from './keys';
 
-export type KeyboardLayoutSource = 'navigator.keyboard' | 'none';
+export type KeyboardLayoutSource = 'navigator.keyboard' | 'user-choice' | 'pressed-keys';
 
 @injectable()
 export class BrowserKeyboardLayoutProvider implements KeyboardLayoutProvider, KeyboardLayoutChangeNotifier {
@@ -30,6 +32,10 @@ export class BrowserKeyboardLayoutProvider implements KeyboardLayoutProvider, Ke
     @inject(ILogger)
     protected readonly logger: ILogger;
 
+    @inject(LocalStorageService)
+    protected readonly storageService: LocalStorageService;
+
+    protected readonly initialized = new Deferred();
     protected readonly nativeLayoutChanged = new Emitter<NativeKeyboardLayout>();
 
     get onDidChangeNativeLayout() {
@@ -37,11 +43,24 @@ export class BrowserKeyboardLayoutProvider implements KeyboardLayoutProvider, Ke
     }
 
     protected readonly tester = new KeyboardTester(loadAllLayouts());
-    protected source: KeyboardLayoutSource = 'none';
-    protected lastSelected?: KeyboardLayoutData;
+    protected source: KeyboardLayoutSource = 'pressed-keys';
+    protected currentLayout: KeyboardLayoutData = DEFAULT_LAYOUT_DATA;
+
+    get allLayoutData() {
+        return this.tester.candidates.slice();
+    }
+
+    get currentLayoutData() {
+        return this.currentLayout;
+    }
+
+    get currentLayoutSource() {
+        return this.source;
+    }
 
     @postConstruct()
-    protected initialize(): void {
+    protected async initialize(): Promise<void> {
+        await this.loadState();
         const keyboard = (navigator as NavigatorExtension).keyboard;
         if (keyboard && keyboard.addEventListener) {
             keyboard.addEventListener('layoutchange', async () => {
@@ -49,22 +68,38 @@ export class BrowserKeyboardLayoutProvider implements KeyboardLayoutProvider, Ke
                 this.nativeLayoutChanged.fire(newLayout);
             });
         }
+        this.initialized.resolve();
     }
 
     async getNativeLayout(): Promise<NativeKeyboardLayout> {
-        const keyboard = (navigator as NavigatorExtension).keyboard;
-        if (keyboard && keyboard.getLayoutMap) {
-            try {
-                const layoutMap = await keyboard.getLayoutMap();
-                this.source = 'navigator.keyboard';
-                this.testLayoutMap(layoutMap);
-            } catch (error) {
-                this.logger.warn('Failed to obtain keyboard layout map.', error);
-            }
+        await this.initialized.promise;
+        if (this.source === 'user-choice') {
+            return this.currentLayout.raw;
         }
-        const layout = this.selectLayout();
-        this.setSelected(layout);
+        const [layout, source] = await this.autodetect();
+        this.setCurrent(layout, source);
         return layout.raw;
+    }
+
+    /**
+     * Set user-chosen keyboard layout data.
+     */
+    async setLayoutData(layout: KeyboardLayoutData | 'autodetect'): Promise<KeyboardLayoutData> {
+        if (layout === 'autodetect') {
+            if (this.source === 'user-choice') {
+                const [newLayout, source] = await this.autodetect();
+                this.setCurrent(newLayout, source);
+                this.nativeLayoutChanged.fire(newLayout.raw);
+                return newLayout;
+            }
+            return this.currentLayout;
+        } else {
+            if (this.source !== 'user-choice' || layout !== this.currentLayout) {
+                this.setCurrent(layout, 'user-choice');
+                this.nativeLayoutChanged.fire(layout.raw);
+            }
+            return layout;
+        }
     }
 
     /**
@@ -74,7 +109,7 @@ export class BrowserKeyboardLayoutProvider implements KeyboardLayoutProvider, Ke
      * change event is fired.
      */
     validateKeyCode(keyCode: KeyCode): void {
-        if (!keyCode.key || !keyCode.character || this.source !== 'none') {
+        if (!keyCode.key || !keyCode.character || this.source !== 'pressed-keys') {
             return;
         }
         const accepted = this.tester.updateScores({
@@ -88,15 +123,35 @@ export class BrowserKeyboardLayoutProvider implements KeyboardLayoutProvider, Ke
             return;
         }
         const layout = this.selectLayout();
-        if (this.lastSelected === undefined || layout !== this.lastSelected && layout !== DEFAULT_LAYOUT_DATA) {
-            this.setSelected(layout);
+        if (layout !== this.currentLayout && layout !== DEFAULT_LAYOUT_DATA) {
+            this.setCurrent(layout, 'pressed-keys');
             this.nativeLayoutChanged.fire(layout.raw);
         }
     }
 
-    protected setSelected(layout: KeyboardLayoutData): void {
-        this.lastSelected = layout;
-        this.logger.info(`Detected keyboard layout: ${layout.name}`);
+    protected setCurrent(layout: KeyboardLayoutData, source: KeyboardLayoutSource): void {
+        this.currentLayout = layout;
+        this.source = source;
+        this.saveState();
+        if (this.tester.inputCount && (source === 'pressed-keys' || source === 'navigator.keyboard')) {
+            const from = source === 'pressed-keys' ? 'pressed keys' : 'browser API';
+            const hardware = layout.hardware === 'mac' ? 'Mac' : 'PC';
+            this.logger.info(`Detected keyboard layout from ${from}: ${layout.name} (${hardware})`);
+        }
+    }
+
+    protected async autodetect(): Promise<[KeyboardLayoutData, KeyboardLayoutSource]> {
+        const keyboard = (navigator as NavigatorExtension).keyboard;
+        if (keyboard && keyboard.getLayoutMap) {
+            try {
+                const layoutMap = await keyboard.getLayoutMap();
+                this.testLayoutMap(layoutMap);
+                return [this.selectLayout(), 'navigator.keyboard'];
+            } catch (error) {
+                this.logger.warn('Failed to obtain keyboard layout map.', error);
+            }
+        }
+        return [this.selectLayout(), 'pressed-keys'];
     }
 
     /**
@@ -132,13 +187,38 @@ export class BrowserKeyboardLayoutProvider implements KeyboardLayoutProvider, Ke
                 topScoringCount++;
             }
         }
-        if (topScoringCount === 1) {
-            return candidates.find((_, i) => scores[i] === topScore)!;
-        }
-        if (matchingOScount === 1) {
+        if (matchingOScount >= 1) {
             return candidates.find((c, i) => scores[i] === topScore && osMatches(c.hardware))!;
         }
+        if (topScoringCount >= 1) {
+            return candidates.find((_, i) => scores[i] === topScore)!;
+        }
         return DEFAULT_LAYOUT_DATA;
+    }
+
+    protected saveState(): Promise<void> {
+        const data: LayoutProviderState = {
+            tester: this.tester.getState(),
+            source: this.source,
+            currentLayout: this.currentLayout !== DEFAULT_LAYOUT_DATA ? getLayoutId(this.currentLayout) : undefined
+        };
+        return this.storageService.setData('keyboard', data);
+    }
+
+    protected async loadState(): Promise<void> {
+        const data = await this.storageService.getData<LayoutProviderState>('keyboard');
+        if (data) {
+            this.tester.setState(data.tester || {});
+            this.source = data.source || 'pressed-keys';
+            if (data.currentLayout) {
+                const layout = this.tester.candidates.find(c => getLayoutId(c) === data.currentLayout);
+                if (layout) {
+                    this.currentLayout = layout;
+                }
+            } else {
+                this.currentLayout = DEFAULT_LAYOUT_DATA;
+            }
+        }
     }
 
 }
@@ -168,12 +248,24 @@ export const DEFAULT_LAYOUT_DATA: KeyboardLayoutData = {
     }
 };
 
+export interface LayoutProviderState {
+    tester?: KeyboardTesterState;
+    source?: KeyboardLayoutSource;
+    currentLayout?: string;
+}
+
 export interface KeyboardTestInput {
     code: string;
     character: string;
     shiftKey?: boolean;
     ctrlKey?: boolean;
     altKey?: boolean;
+}
+
+export interface KeyboardTesterState {
+    scores?: { [id: string]: number };
+    topScore?: number;
+    testedInputs?: { [key: string]: string }
 }
 
 /**
@@ -187,6 +279,10 @@ export class KeyboardTester {
     topScore: number = 0;
 
     private readonly testedInputs = new Map<string, string>();
+
+    get inputCount() {
+        return this.testedInputs.size;
+    }
 
     constructor(readonly candidates: KeyboardLayoutData[]) {
         this.scores = this.candidates.map(() => 0);
@@ -243,6 +339,47 @@ export class KeyboardTester {
         }
     }
 
+    getState(): KeyboardTesterState {
+        const scores: { [id: string]: number } = {};
+        for (let i = 0; i < this.scores.length; i++) {
+            scores[getLayoutId(this.candidates[i])] = this.scores[i];
+        }
+        const testedInputs: { [key: string]: string } = {};
+        for (const [key, character] of this.testedInputs.entries()) {
+            testedInputs[key] = character;
+        }
+        return {
+            scores,
+            topScore: this.topScore,
+            testedInputs
+        };
+    }
+
+    setState(state: KeyboardTesterState): void {
+        this.reset();
+        if (state.scores) {
+            const layoutIds = this.candidates.map(getLayoutId);
+            for (const id in state.scores) {
+                if (state.scores.hasOwnProperty(id)) {
+                    const index = layoutIds.indexOf(id);
+                    if (index > 0) {
+                        this.scores[index] = state.scores[id];
+                    }
+                }
+            }
+        }
+        if (state.topScore) {
+            this.topScore = state.topScore;
+        }
+        if (state.testedInputs) {
+            for (const key in state.testedInputs) {
+                if (state.testedInputs.hasOwnProperty(key)) {
+                    this.testedInputs.set(key, state.testedInputs[key]);
+                }
+            }
+        }
+    }
+
 }
 
 /**
@@ -259,10 +396,22 @@ interface Keyboard {
 
 type KeyboardLayoutMap = Map<string, string>;
 
+function getLayoutId(layout: KeyboardLayoutData): string {
+    return `${layout.language}-${layout.name.replace(' ', '_')}-${layout.hardware}`;
+}
+
+/**
+ * Keyboard layout files are expexted to have the following name scheme:
+ *     `language-name-hardware.json`
+ *
+ * - `language`: A language subtag according to IETF BCP 47
+ * - `name`:     Display name of the keyboard layout (without dashes)
+ * - `hardware`: `pc` or `mac`
+ */
 function loadLayout(fileName: string): KeyboardLayoutData {
     const [language, name, hardware] = fileName.split('-');
     return {
-        name: `${name} (${hardware === 'mac' ? 'Mac' : 'PC'})`,
+        name: name.replace('_', ' '),
         hardware: hardware as 'pc' | 'mac',
         language,
         raw: require('../../../src/common/keyboard/layouts/' + fileName + '.json')
@@ -270,12 +419,57 @@ function loadLayout(fileName: string): KeyboardLayoutData {
 }
 
 function loadAllLayouts(): KeyboardLayoutData[] {
+    // The order of keyboard layouts is relevant for autodetection. Layouts with
+    // lower index have a higher chance of being selected.
+    // The current ordering approach is to sort by estimated number of developers
+    // in the respective country (taken from the Stack Overflow Developer Survey),
+    // but keeping all layouts of the same language together.
     return [
-        'de-German-pc',
-        'de-German-mac',
         'en-US-pc',
         'en-US-mac',
+        'en-Dvorak-pc',
+        'en-Dvorak-mac',
+        'en-Dvorak_Lefthanded-pc',
+        'en-Dvorak_Lefthanded-mac',
+        'en-Dvorak_Righthanded-pc',
+        'en-Dvorak_Righthanded-mac',
+        'en-Colemak-mac',
+        'en-British-pc',
+        'en-British-mac',
+        'de-German-pc',
+        'de-German-mac',
+        'de-Swiss_German-pc',
+        'de-Swiss_German-mac',
         'fr-French-pc',
-        'fr-French-mac'
+        'fr-French-mac',
+        'fr-Canadian_French-pc',
+        'fr-Canadian_French-mac',
+        'fr-Swiss_French-pc',
+        'fr-Swiss_French-mac',
+        'pt-Portuguese-pc',
+        'pt-Portuguese-mac',
+        'pt-Brazilian-mac',
+        'pl-Polish-pc',
+        'pl-Polish-mac',
+        'nl-Dutch-pc',
+        'nl-Dutch-mac',
+        'es-Spanish-pc',
+        'es-Spanish-mac',
+        'it-Italian-pc',
+        'it-Italian-mac',
+        'sv-Swedish-pc',
+        'sv-Swedish-mac',
+        'tr-Turkish_Q-pc',
+        'tr-Turkish_Q-mac',
+        'cs-Czech-pc',
+        'cs-Czech-mac',
+        'ro-Romanian-pc',
+        'ro-Romanian-mac',
+        'da-Danish-pc',
+        'da-Danish-mac',
+        'nb-Norwegian-pc',
+        'nb-Norwegian-mac',
+        'hu-Hungarian-pc',
+        'hu-Hungarian-mac'
     ].map(loadLayout);
 }